# backend/src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "FT_Transcendence API",
    "description": "API Documentation for FT_Transcendence Project"
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      },
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
...
  },
  "paths": {
...
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresh access token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get all users",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search users by username"
          }
        ],
        "responses": {
          "200": {
            "description": "List of users"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/friends": {
      "get": {
        "tags": [
          "Friends"
        ],
        "summary": "Get friends list",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of friends"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/public/users": {
      "get": {
        "tags": [
          "Public API"
        ],
        "summary": "Get all users",
        "security": [
          {
            "api_key": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of users"
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Too many requests"
          }
        }
      },
      "post": {
        "tags": [
          "Public API"
        ],
        "summary": "Create a new user record",
        "security": [
          {
            "api_key": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 10
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "avatar_url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "level": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "online",
                      "offline",
                      "in_game"
                    ]
                  }
                },
                "required": [
                  "username",
                  "email"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Username or email exists"
          },
          "429": {
            "description": "Too many requests"
          }
        }
      }
    },
...
  }
}